// src/app/modules/admin/cutting-lists/models/productSummaryDetailsDTO.ts
export interface ProductSummaryDetailsDTO {
  priority: string;
  productDateWorkBegan: string;
  productDateWorkCompleted: string;
  productDateLastWorkedOn: string;
  profile: string;
  colorId: number;
  gaugeId: number;
  widthId: number;
  width: number;
  gauge: number;
  color: string;
  productId: number;
  targetDate: string;
  productStatus: string;
  planName: string;
  canInvoice: boolean;
  profileId: number;
  quoteDateAccepted: string;
  clientName: string;
  quoteStatus: string;
  completedLength: number;
  scheduledLength: number;
}